Refactoring: remove seldom-used BaseAction class
